% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtag.R
\name{autoplot.qtag.long}
\alias{autoplot.qtag.long}
\alias{autoplot.qtag.wide}
\title{Autoplot a qualifier/tag structure}
\usage{
\method{autoplot}{qtag.long}(x, xvar, yvar, facets, ...)

\method{autoplot}{qtag.wide}(x, ...)
}
\arguments{
\item{x}{A \link{qtag} object}

\item{xvar}{Column to be used on the x-axis}

\item{yvar}{Column to be used on the y-axis}

\item{facets}{Column to be used as facetting variable}

\item{...}{Passed on to \code{aes()}}
}
\value{
A ggplot object
}
\description{
Autoplot a qualifier/tag structure
}
\examples{
data(pocmaj)
autoplot(qtag(pocmaj, qualifiers=c("core", "depth")))

}

