% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_segments.R
\name{linseg.classify}
\alias{linseg.classify}
\title{Classify depths from linear segments}
\usage{
linseg.classify(linseg, depth, param = "name", preference = c("upper",
  "lower"))
}
\arguments{
\item{linseg}{A \code{linseg} object (see \link{linseg})}

\item{depth}{A vector of depths to classify}

\item{preference}{Preference for upper or lower unit should a depth match more than one unit.}

\item{value}{The parameter to return (defaults to 'name')}
}
\value{
A character or factor
}
\description{
Classify depths from linear segments
}
\examples{
data(lithozones_example)
linseg.classify(lithozones_example, 5)
linseg.classify(lithozones_example, 5, preference="lower")
linseg.classify(lithozones_example, seq(0, 26, 2))
linseg.classify(lithozones_example, seq(0, 26, 2), param="gs") # gives grain size for each depth
linseg.classify(lithozones_example, seq(0, 26, 2), param="gss") # gives NAs

}

