% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_segments.R
\name{linseg}
\alias{linseg}
\title{Create a linear segment object}
\usage{
linseg(breaks, names, name = NA, ...)
}
\arguments{
\item{breaks}{A vector of segment breaks (must be at least length 2)}

\item{names}{A vector of names for the segments}

\item{name}{The name associated with the linear segments (optional)}

\item{...}{Additional vectors of length equal to length of \code{names} that correspond
to additional parameters (e.g. lithology, grain size)}
}
\value{
A \code{linseg} object.
}
\description{
Creates a linear segment object from minimal information.
}
\examples{
ls <- linseg(c(0, 1, 2, 3))
plot(ls)
ls <- linseg(c(0, 1, 2, 3), names=c("Unit 1", "Unit 2", "Unit 3"),
             grain_size=c("fg", "mg", "cg"))
plot(ls)

}

