% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lithozones_example}
\alias{lithozones_example}
\title{Lithozones Example Dataset}
\format{A \code{linseg} object for use with the appropriate linear segiment
  functions.}
\usage{
lithozones_example
}
\description{
Base lithozones example dataset.
}
\keyword{datasets}

