% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_segments.R
\name{plot.linseg}
\alias{plot.linseg}
\title{Plot a linear segment object}
\usage{
\method{plot}{linseg}(x, ..., label_by = "name", reversed = TRUE,
  ylab = NA, xlab = NA, sub = NA, xright = 1, xleft = 0, xlim = c(0,
  4), ylim = NULL, col = NULL, label.col = "black", label.margin = 0.1)
}
\arguments{
\item{x}{A \code{linseg} object}

\item{...}{Formatting options to be passed on to \code{rect}.}

\item{label_by}{A parameter with which to label segments or \code{NULL} for no label.}

\item{reversed}{Logical describing if the y axis should be reversed.}

\item{ylab}{Label for the Y axis}

\item{xlab}{Label for the X axis}

\item{sub}{Subtitle for the plot}

\item{xright}{Right X value for plot rectangles (can be a vector)}

\item{xleft}{Left X value for plot rectangles (can be a vector)}

\item{xlim}{vector of length 2 for plot X range.}

\item{ylim}{vector of length 2 for plot Y range (overrides \code{reversed} parameter)}

\item{col}{Colour vector for segments}

\item{label.col}{Colour vector for labels}

\item{label.margin}{Margin vector for labels}
}
\description{
Provide a visual of a \code{linseg} object.
}
\examples{
data(lithozones_example)
plot(lithozones_example)

}

