% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtag.R
\name{qtag}
\alias{qtag}
\title{Construct a qualifier/tag structure}
\usage{
qtag(df, values, qualifiers, tags, summarised = FALSE, quiet = FALSE)
}
\arguments{
\item{df}{A \code{data.frame} or similar object}

\item{values}{Column names containing the values of interest}

\item{qualifiers}{Column names of qualifying values}

\item{tags}{Column names of tag values}

\item{summarised}{Pass \code{TRUE} for pre-summarised data.}

\item{quiet}{Use \code{quiet=TRUE} to suppress error messages}
}
\value{
An object of type \code{qtag}, which is essentially the unchanged
  input with qualifiers, values, and tags information attached.
}
\description{
Construct a qualifier/tag structure
}
\examples{
data("pocmaj")
pocmaj <- qtag(pocmaj, qualifiers = c("core", "depth"))
long(pocmaj)
aggregate(pocmaj)
aggregate(long(pocmaj))

}

