% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtag.R
\name{qualifiers}
\alias{qualifiers}
\title{Extract value qualifier names from a qualifier/tag structure}
\usage{
qualifiers(x)
}
\arguments{
\item{x}{A \link{qtag} object}
}
\value{
A vector of qualifier column names
}
\description{
Extract value qualifier names from a qualifier/tag structure
}
\examples{
data(pocmaj)
pocmaj <- qtag(pocmaj, qualifiers=c("core", "depth"))
qualifiers(pocmaj)

}

