% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_segments.R
\name{linseg.param}
\alias{linseg.param}
\title{Get param values for linear segments}
\usage{
linseg.param(linseg, name, param)
}
\arguments{
\item{linseg}{A \code{linseg} object (see \link{linseg})}

\item{param}{The parameter name to return}
}
\value{
A character vector of param values, or a \code{list} if some params have
        multiple values.
}
\description{
Get param values for linear segments
}
\examples{
data(lithozones_example)
linseg.param(lithozones_example, "Unit 1", "gs")

}

